Implemented Reviews and Star Rating feature, + a simple design for them. - Added an 'ArtistReview' model specific for artists. - Used the 'ProductReview' and 'ArtistReview' models to get reviews from database and display them in the comments sections (in product details and artist details templates).
- Added the script needed to use jquery in base.html template, but haven't used it yet.
NOTE : The script is supposed to display reviews without having to reload the page, but I haven't implemented that part yet.
BIG NOTE : The 'reviews' views are rendering the same template in which the review is added; I had to duplicate parts of code from artist/product - detail views in order to make it work; all in the sake of refreshing the page so that  it displays the new review. There's also a problem with this : If you refresh the page, it'll ask you to confirm refreshing given that the browser will repeat any actions he's done before (which means the review added will be duplicated). I'll fix this later.