More modifications to the home page, added most of the missing pages, added Orders chart in dashboard page.
Home page :
- Testimonials and Ratings sections in home page are now working :
--> Had to create a form for each, here's why :
I tried using the same structure of form for artist and products review + some modifications, but it just didn't work. Multiple types of errors, such as multivaluedict error and others, also it wasn't possible for the user to enter either of them (well either that or I'm a noob :P)

Pages :

- Pages such as services, contact, etc, are added and functional.

- Delaying animations until user reaches them is now functional.
--> Considering adding slideInFromBottom animation to everything and not just the current elements.

Dashboard :
Using Chartjs, I added a chart that'll represent the amount of user's orders per month : it shows how many orders the user ordered in which month.